Initial styling pass
