+++ server/handlebars/public/style.css
@@ -1,0 +1,93 @@
+#main {
+    width: 700px;
+}
+
+#candidate-title {
+    border: 1.5px solid #FFC4A2;
+    border-top-left-radius: 3px;
+    border-top-right-radius: 3px;
+    background-color: #FFD8C1;
+    padding: .5em .5em .5em 1em;
+}
+
+#candidate-title > span {
+    font-family: Arial, Helvetica, sans-serif;
+    font-size: 12px;
+    font-weight: bold;
+    color: white;
+    background-color: #C25401;
+    padding: 5px;
+    border-radius: 15px;
+}
+
+#candidate-content-container {
+    border-left: 1.5px solid #D5D8DD;
+    border-right: 1.5px solid #D5D8DD;
+    border-bottom: 1.5px solid #D5D8DD;
+    border-bottom-left-radius: 3px;
+    border-bottom-right-radius: 3px;
+    padding: 1em .5em .5em 1em;
+    font-family: Arial, Helvetica, sans-serif;
+    font-size: 14px;
+    margin-bottom: 1em;
+}
+
+#candidate-content-container > ol > li:not(:last-child) {
+    margin-bottom: 3px;
+}
+
+.none {
+    display: block;
+    color: #D5D8DD;
+    font-style: italic;
+    text-align: center;
+}
+
+table {
+    border-collapse: collapse;
+    border-top: 1.5px solid #D5D8DD;
+    width: 100%;
+    margin-bottom: 1em;
+  }
+
+table td, th {
+    padding: 1rem;
+    border: 1.5px solid #D5D8DD; 
+}
+
+th, td {
+    font-family: Arial, Helvetica, sans-serif;
+    padding: 15px;
+    border: 1.5px solid #D5D8DD;
+}
+
+th {
+    background-color: #F2F2F2;
+}
+
+table tbody tr th {
+    text-align: left;
+}
+
+table tbody tr td {
+    padding: 20px;
+}
+
+.meter {
+  box-sizing: content-box;
+  height: 15px; /* Can be anything */
+  position: relative;
+  background: #D5D8DD;
+  border-radius: 25px;
+  margin-bottom: 10px;
+  overflow: hidden;
+}
+
+.meter > span {
+    display: block;
+    height: 100%;
+    float: left;
+    background-color: #175A6A;
+    border-radius: 25px;
+    position: relative;
+  }
