+++ server/handlebars/views/layouts/index.hbs
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />  
     <title>ARIA-AT Report</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0, viewport-fit=cover" />
-    <link rel="stylesheet" type="text/css" href="./style.css" />
+    <link rel="stylesheet" type="text/css" href="/embed/style.css" />
   </head>
   <body>
     {{{body}}}
