+++ server/handlebars/views/main.hbs
@@ -1,16 +1,16 @@
-<div>
+<div id="main">
     {{#if (isCandidate status)}}
     <div id="embded-report-status-container">
-    <div><span>Candidate Report</span></div>
-    <div>
-        The information in this report generated from candidate tests.
-        Candidate aria-at tests are in review by assistive technology
-        developer and lack consensus regarding:
-        <ol>
-            <li>applicability and validity of the tests, and</li>
-            <li>accuracy of test results.</li>
-        </ol>
-    </div>
+        <div id="candidate-title"><span>Candidate Report</span></div>
+        <div id="candidate-content-container">
+            The information in this report generated from candidate tests.
+            Candidate aria-at tests are in review by assistive technology
+            developer and lack consensus regarding:
+            <ol>
+                <li>applicability and validity of the tests, and</li>
+                <li>accuracy of test results.</li>
+            </ol>
+        </div>
     </div>
     {{/if}}
     <table id="embed-report-table">
@@ -29,23 +29,38 @@
                     <th><b>{{this.at.name}}</b></th>
                     {{#if (isInAllBrowsers "Chrome" @../../this)}}
                         {{#if (isBrowser this.browser.name "Chrome") }}
-                        <td><span classname="progress50"></span><b>{{this.metrics.supportPercent}}%</b> of passing assertions</b></td>
+                        <td>
+                            <div class="meter">
+                                <span style="width: {{this.metrics.supportPercent}}%;"></span>
+                            </div>
+                            <b>{{this.metrics.supportPercent}}%</b> of passing assertions</b>
+                        </td>
                         {{else}}
-                        <td>None</td>
+                        <td><span class="none">None</span></td>
                         {{/if}} 
                     {{/if}} 
                     {{#if (isInAllBrowsers "Safari" @../../this)}}
                         {{#if (isBrowser this.browser.name "Safari") }}
-                        <td><span classname="progress50"></span><b>{{this.metrics.supportPercent}}%</b> of passing assertions</b></td>
+                        <td>
+                            <div class="meter">
+                                <span style="width: {{this.metrics.supportPercent}}%;"></span>
+                            </div>
+                            <b>{{this.metrics.supportPercent}}%</b> of passing assertions</b>
+                        </td>
                         {{else}}
-                        <td>None</td>
+                        <td><span class="none">None</span></td>
                         {{/if}} 
                     {{/if}} 
                     {{#if (isInAllBrowsers "Firefox" @../../this)}}
                         {{#if (isBrowser this.browser.name "Firefox") }}
-                        <td><span classname="progress50"></span><b>{{this.metrics.supportPercent}}%</b> of passing assertions</b></td>
+                        <td>
+                            <div class="meter">
+                                <span style="width: {{this.metrics.supportPercent}}%;"></span>
+                            </div>
+                            <b>{{this.metrics.supportPercent}}%</b> of passing assertions</b>
+                        </td>
                         {{else}}
-                        <td>None</td>
+                        <td><span class="none">None</span></td>
                         {{/if}} 
                     {{/if}}          
                 </tr>
